use global fetch directly

+++ scripts/build-sdl.mjs
@@ -1,21 +1,22 @@
 import Fs from 'node:fs'
 import Path from 'node:path'
-import { once } from 'node:events'
+import Stream from 'node:stream'
 import C from './util/common.js'
-import { fetch } from './util/fetch.js'
 import * as Tar from 'tar'
 
 const url = `https://github.com/${C.sdl.owner}/${C.sdl.repo}/archive/refs/tags/v${C.sdl.version}.tar.gz`
 
 console.log("fetch", url)
 const response = await fetch(url)
+if (!response.ok) { throw new Error(`bad status code ${response.status}`) }
 
 console.log("unpack to", C.dir.sdl)
 await Fs.promises.rm(C.dir.sdl, { recursive: true }).catch(() => {})
 await Fs.promises.mkdir(C.dir.sdl, { recursive: true })
-const tar = Tar.extract({ gzip: true, C: C.dir.sdl })
-response.stream().pipe(tar)
-await once(tar, 'finish')
+await Stream.promises.pipeline(
+	Stream.Readable.fromWeb(response.body),
+	Tar.extract({ gzip: true, C: C.dir.sdl }),
+)
 
 const sdlRoot = Path.join(C.dir.sdl, `${C.sdl.repo}-${C.sdl.version}`)
 await import(Path.join(sdlRoot, '/scripts/build.mjs'))
